Revert: move proof_size back off FRIParams
Reverts #2244 at jimpo's request: "Unnecessary coupling, at odds with the
single-responsibility principle." So proof_size returns to a free
function in its own module, taking the params and the scheme as
arguments.

The doc comment #2244 rewrote is kept. It reads one clause per line and
splits the two channels into their own lists, rather than wrapping a
sentence across lines inside one bullet.

Two call sites landed after #2244 and use the method form, so they
convert too: the proof_size_at_rate helper in fri::common's tests, and
the FRI-vs-Ligerito size comparison in ligerito::size_estimation.
